+++ src/fasta.h
@@ -13,5 +13,5 @@ int fasta_read(FILE *stream, const char *accept, char **id,
                size_t *id_size, char **seq, size_t *seq_size);
 
 void fasta_write(FILE *stream, const char *id, const char *comment,
-                 const char *seq, size_t width);
+                 const char *seq, int width);
 #endif
